+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+0.3.3
+*****
+* Added ``PathType``, an annotation for `path-like <https://docs.python.org/3/glossary.html#term-path-like-object>`_ objects.
+* Added the ``copy`` argument to ``as_nd_array()``.
+
+
 0.3.2
 *****
 * Fixed a bug with ``split_dict()``.